+++ src/components/MenuBar.vue
@@ -124,6 +124,8 @@ const onAppMenu = (event: Event) => {
             [{ label: t('tray.menu.installUpdate'), onClick: () => window.api.update.apply() }] :
             [{ label: t('menu.app.checkForUpdates'), onClick: () => window.api.update.check() }]
           ),
+          { label: t('menu.file.backupExport'), divided: 'up', onClick: () => window.api.backup.export() },
+          { label: t('menu.file.backupImport'), onClick: () => window.api.backup.import() },
           { label: t('menu.file.closeWindow'), divided: 'up', onClick: () => window.api.main.close() },
         ]
       },